Move TypeVariableChecker to mypy.semanal_pass3
